Combine LivingFoodEvent & LivingFoodEffectEvent
