+++ src/main/java/net/minecraftforge/event/entity/living/LivingEatEvent.java
@@ -5,8 +5,13 @@
 
 package net.minecraftforge.event.entity.living;
 
+import java.util.List;
+
 import org.jetbrains.annotations.Nullable;
 
+import com.mojang.datafixers.util.Pair;
+
+import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.food.FoodData;
 import net.minecraft.world.food.FoodProperties;
@@ -20,7 +25,7 @@ import net.minecraftforge.eventbus.api.Cancelable;
  * <br>
  * This event is {@link Cancelable}.<br>
  * If this event is canceled, the food item is consumed but no changes will be
- * made to the hunger and saturation.<br>
+ * made to the hunger and saturation and no effects will be applied.<br>
  * <br>
  * This event is fired on the {@link MinecraftForge#EVENT_BUS}.
  */
@@ -30,24 +35,28 @@ public class LivingEatEvent extends LivingEvent
     private int foodAmount;
     private float saturationAmount;
     private final ItemStack foodItem;
+    private final List<Pair<MobEffectInstance, Float>> effectPairs;
 
-    private LivingEatEvent(@Nullable LivingEntity entity, int foodAmount, float saturationAmount,
-            ItemStack foodItem)
+    public LivingEatEvent(@Nullable LivingEntity entity, ItemStack foodItem)
     {
         super(entity);
-        this.foodAmount = foodAmount;
-        this.saturationAmount = saturationAmount;
         this.foodItem = foodItem;
+        if(foodItem != null)
+        {
+            FoodProperties prop = foodItem.getFoodProperties(entity);
+            this.foodAmount = prop.getNutrition();
+            this.saturationAmount = prop.getSaturationModifier();
+            this.effectPairs = prop.getEffects();
+        }
+        else
+            this.effectPairs = null;
     }
 
     public LivingEatEvent(@Nullable LivingEntity entity, int foodAmount, float saturationAmount)
     {
-        this(entity, foodAmount, saturationAmount, null);
-    }
-
-    public LivingEatEvent(@Nullable LivingEntity entity, ItemStack foodItem, FoodProperties prop)
-    {
-        this(entity, prop.getNutrition(), prop.getSaturationModifier(), foodItem);
+        this(entity, null);
+        this.foodAmount = foodAmount;
+        this.saturationAmount = saturationAmount;
     }
 
     /**
@@ -90,5 +99,24 @@ public class LivingEatEvent extends LivingEvent
     public ItemStack getFoodItem()
     {
         return foodItem;
+    }
+
+    /**
+     * can be modified to change the applied effects and probability
+     * @return the effect-probability pair of {@link #foodItem}
+     */
+    public List<Pair<MobEffectInstance, Float>> getEffects()
+    {
+        return effectPairs;
+    }
+    
+    /**
+     * adds a new {@link MobEffectInstance} and probability to apply
+     * @param effect {@link MobEffectInstance} to apply
+     * @param probability probability of application
+     */
+    public void addEffect(MobEffectInstance effect, float probability)
+    {
+        this.effectPairs.add(new Pair<MobEffectInstance, Float>(effect, probability));
     }
 }
