+++ src/test/java/net/minecraftforge/debug/entity/living/LivingFoodEventTest.java
@@ -16,7 +16,6 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.living.LivingEatEvent;
-import net.minecraftforge.event.entity.living.LivingFoodEffectEvent;
 import net.minecraftforge.fml.common.Mod;
 
 @Mod("living_food_event_test")
@@ -27,21 +26,24 @@ public class LivingFoodEventTest
     public LivingFoodEventTest()
     {
         MinecraftForge.EVENT_BUS.addListener(this::onLivingEat);
-        MinecraftForge.EVENT_BUS.addListener(this::onLivingFoodEffect);
     }
 
     // Tests the LivingEatEvent
     // Logs the items a LivingEntity is about to eat
-    // Cancels the event if it is Steak
+    // Cancels the event if it is spider eye
     // Changes to 1 nutrition and 0 saturation for anything else
+    //
+    // Entirely removes potion effects from eating a golden apple
+    // Add a regeneration 4 effect to pufferfishes
     public void onLivingEat(LivingEatEvent event)
     {
         LOGGER.info("{} is about to eat {} with {} nutrition and {} saturation", event.getEntity().getType(),
                 event.getFoodItem().getItem(), event.getFoodAmount(), event.getSaturationAmount());
-        if (event.getFoodItem().getItem() == Items.COOKED_BEEF)
+        Item item = event.getFoodItem().getItem();
+        if (item == Items.SPIDER_EYE)
         {
             event.setCanceled(true);
-            LOGGER.info("Canceled eating Steak");
+            LOGGER.info("Canceled eating spider eye");
         }
         else
         {
@@ -49,14 +51,7 @@ public class LivingFoodEventTest
             event.setSaturationAmount(0);
             LOGGER.info("Changed to 1 nutrition and 0 saturation");
         }
-    }
-
-    // Tests the LivingFoodEffectEvent
-    // Entirely removes potion effects from eating a golden apple
-    // Add a regeneration 4 effect to pufferfishes
-    public void onLivingFoodEffect(LivingFoodEffectEvent event)
-    {
-        Item item = event.getFoodItem().getItem();
+        
         if (item == Items.GOLDEN_APPLE)
         {
             event.getEffects().clear();
@@ -64,8 +59,7 @@ public class LivingFoodEventTest
         }
         else if (item == Items.PUFFERFISH)
         {
-            event.getEffects().add(
-                    new Pair<MobEffectInstance, Float>(new MobEffectInstance(MobEffects.REGENERATION, 100, 3), 1f));
+            event.getEffects().add(new Pair<MobEffectInstance, Float>(new MobEffectInstance(MobEffects.REGENERATION, 100, 3), 1f));
             LOGGER.info("Added Regeneration to PufferFish");
         }
     }
